requirements/base.txt
asgiref==3.5.2
Django==3.2.16
pytz==2022.6
sqlparse==0.4.3
django-jazzmin==2.6.0
pre-commit==2.20.0
pylint==2.13.7
pylint-django==2.5.3
django-money==3.0.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
mypy==0.991
django-import-export==3.0.2
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
python-decouple==3.7
